guardrails-demo/requirements.txt
openai 
guardrails-ai 
python-dotenv 
langchain 
rapidfuzz 
pandas